feat: Tournament stats with SwiftUI
